+++ figures/cercle.cc
@@ -4,5 +4,5 @@ Cercle::Cercle(const Cercle & r)
     : Forme_simple_coord::Forme_simple_coord(r.positionX,r.positionY), rayon(r.rayon) {}
 
 
-Cercle::Cercle(double x, double y, double rayon)
+Cercle::Cercle(int x, int y, int rayon)
     : Forme_simple_coord::Forme_simple_coord(x,y), rayon(rayon) {}
