+++ figures/ligne.hh
@@ -4,12 +4,13 @@
 
 class Ligne : public Forme {
 public:
-    double positionX1;
-    double positionY1;
-    double positionX2;
-    double positionY2;
+    int positionX1;
+    int positionY1;
+    int positionX2;
+    int positionY2;
 
-    Ligne(double x1, double y1, double x2, double y2);
+    Ligne(int x1, int y1, int x2, int y2);
     Ligne(const Ligne & r);
     ~Ligne() = default;
+
 };
