+++ figures/rectangle.hh
@@ -1,19 +1,19 @@
 #pragma once
 #include "forme.hh"
 
-
 class Rectangle : public Forme {
 public:
-    double positionX1;
-    double positionY1;
-    double positionX2;
-    double positionY2;
-    double positionX3;
-    double positionY3;
-    double positionX4;
-    double positionY4;
+    int positionX1;
+    int positionY1;
+    int positionX2;
+    int positionY2;
+    int positionX3;
+    int positionY3;
+    int positionX4;
+    int positionY4;
 
-    Rectangle(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4);
+    Rectangle(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4);
     Rectangle(const Rectangle & r);
     ~Rectangle() = default;
+
 };
